Switch to Gradle 1.11 & simplified authorizer factory
Gradle 1.11 - I'm using this for our 'universal' build script so I want
to be consistent everywhere. https://github.com/couchbase/couchbase-lite-java-core/issues/179

I also simplified the AuthorizerFactoryManager so it gives each replication
request to the AuthorizerFactory's and let's them figure out if they want it.
Before I required the request to be remote and to have an auth section and
that may or may not be right depending on what the authorizer is up to. In
our case we actually want to trigger off the presence of a HttpKey URL.