Added new getReplicator method for the replication manager
We have some special logic in our listener based replication handler
that deals with things like establishing outgoing mutual SSL auth
connections and soon Tor as well. To make it easy for the replication
manager to leverage this capability we created a new getReplicator
method that just takes ReplicatorArguments as its input.